move create_image to utils
